added and completed faculty section (almost)

+++ remove_application.php
@@ -9,9 +9,19 @@ if(!isset($_SESSION['ldap_id']))
     die("Please login first");
 }
 
+function date_not_passed($date)
+{
+    $todays_date = date('Y-m-d');
+    if(strtotime($date) > strtotime($todays_date))
+    {
+        return true;
+    }
+    return false;
+}
+
 foreach($_POST as $course_code => $remove)
 {
-    if($remove == "Remove")
+    if($remove == "Remove" && date_not_passed($_POST['deadline']))
     {
         $query = "DELETE FROM student_applications WHERE course_code='".$course_code."'";
         if(mysqli_query($conn, $query))
@@ -24,7 +34,14 @@ foreach($_POST as $course_code => $remove)
             echo "There was some error while deleting the entry. Please contact Aman Virani - 9821212128";
             die();
         }
-    }   
+    }
+    else
+    {
+        if(!date_not_passed($_POST['deadline']))
+        {
+            die("Date of removal of application has passed!");
+        }
+    }
 }
